Update TOC fields for 11.1.0
Patch 11.1.0 is adding two new TOC fields for use in the improved addon
list - Category and Group.

Technically both of these fields support localized variants, but of the
two only Category is ever shown in the UI and Group is just an arbitrary
"tag" of sorts that isn't displayed. As such, I'll pretend that `##
Group-deDE: Foo` isn't a thing and won't consider it a valid token.